generator/analysis: add testcases for RunningTaskAnalysis
The RunningTask analysis is quite hard to work on, therefore the
analysis got a verifier framework, to test various results of the
analysis. The verify script is specific for the application, and can
specify checks before and after each analysis run. For the first
application such an verify script is already in place.

Change-Id: I1cadb5acf1cde0acf0187a5c23c39ec313d41125