generator/analyze: add support for ISR2s
ISR2 are handled in the following way: When a non-periodic isr event
is triggered, a new control flow simulation for exaclty that ISR is
started with the a the state of the computation block as input. The
result of that control flow analysis is the return state of the
computation block.

Change-Id: I1493b903301847961989dd015fb169f5e4db67ba