arch/i386: workarounds for FAIL* import-trace
FAIL* import-trace is broken and does not assign symbols to sections.
As a fall-back, the .text section is interpreted as a long chain of
instructions. When read-only data is contained in the .text section
(the IDT, string constants, ...) this data is interpreted as instructions
as well. If the last bytes of such data form an instruction spanning over
the end of the data this will prevent real code follwing the data to be
identified. As a result, valid instructions will be rejected as "not
found" by import-trace.

As a workaround, interrupt handlers are now padded with int3 instructions
(should have been the case anyway) and the linker scripts adds 16 bytes
of int3 after each (potential) data section.

Change-Id: I39848e78a20934f0024595a07096708d5ac8fd2d