generator: instanciate TaskList from the application knowledge
The TaskList is almost fully depended on the task structure of the
application, therefore we can template it in the generator. Therefore
a code templating engine was put in place. os/scheduler/scheduler.h.in
is now used by the rule set for encoded systems. It is not fully operational yet.

Change-Id: I39a8c6fb249ea07792dd58867f2ce20b1a366d57