os: split Task structure in an arch depended/independed part
The stackpointer that is used for x86 during the dispatch is highly
architecutre depended, therefore we're using an architecture specific
TCB structure, that is used to manage that data. arch::TCB is now
generated as a part of the architecture specific part and passed as a
(const!) reference to os::scheduler::Task.

generated-system.cc was fixed NOTDUERFTIG. It should dissappear soon.

Change-Id: If8af895c7f83d2dd78a441d1dd88c9fe285333e1