generator: implement SetRelAlarm
SetRelAlarm is the first systemcall, that really needs arguments from
userspace, that are not always deducible from the task activation
scheme. We encode the values on the userstack and pass a pointer to the
encoded values to the kernelspace. The kernelspace only reads from those
pointers, therefore we can expect no corrupting writes in outerspace.

Change-Id: I908722b00581aacd20a5cd054939c74cbc534cef