os/scheduler: remove scheduler lock
The encoded boolean scheduler lock is removed and replaced
by raising non-preemptive task priorities to RES_SCHEDULER
as suggested in the OSEK spec.

RES_SCHEDULER is determined by the generator to be higher
than the maximum task priority.

As an optimization, rescheduling (calling TaskList::head) is
skipped when current_prio == RES_SCHEDULER.

Change-Id: Id2021a06e6747ed7490b8a1a3cfac1a944ea1545