generator/analysis: new pass DynamicPriorityAnalysis
The dynamic priority pass calculates a dynamic priority for each atomic
basic block. When no resources are used the dynamic priority of each ABB
is the static priority of the subtask it belongs to. When a resource is
taken the static priority of the resource flows on the local task
control graph until the ReleaseResource is done.

We do one non-conforming constraint here: It must be at every ABB
exactly sure which resources are taken and which resources are not
taken. This prohibits constructs like

if (X)
   GetResource(A)

...

if (X)
   ReleaseResource(A)

Change-Id: I41f3caf8710bc7c8a8d29770f613d2fbd6bde153