generator: support multiple Resources
For supporting multiple Resources, and not only RES_SCHEDULER, another
pass was implemented. The PrioritySpreadingPass, which reallocates the
the static priorities of all subtasks and resources. A resource now has
max([task.static_priority]) + 1 and no other task has that priority. By
this the scheduling is strictly priority based again, since owning an
priority is better than being the highest prior task in a resource task set.

Change-Id: I0b8ced1904a4ef723707474f806f8f1d02f2296d