generator/analysis: bugfix in scheduling
When a subtask is ready and has more than one continuation point with
different dynamic priorities, we cannot be sure anymore that always the
higher prior block is taken.

The problem can be illustrated like this. Imagine you activate a higher
prior task, once with an resource taken (prio 6) and once without the
resource (prio 2). In both cases you're scheduled. Then your task is
surely running, but it is not sure that you return that you return
immediately to the successor of the higher prior task.

Change-Id: I74272c8b9570619a3863bc65fb18c32c865d38e3