os: do not use/store D in Encoded_Static
So far, the dynamic signature (D) is not used by the system, but is
stored in memory and can be corrupted. Also, Encoded_Static variables
are larger than 32 Bit because of the added D value.
Now, Encoded_Static discards all updates to D and returns 0 for all reads.
An Encoded_Dynamic subclass is added which provides the old implementation.
In the future, another subclass might be useful which uses a reference to
some D which can then be shared by multiple encoded values.

Change-Id: I8d536867f440fd66bad2bd8565e1197673574e93