generator: add --specialize for specialized system calls
The SpecializedSystemCalls backend exploits the RunningTask analysis to
generate very specific code for exactly one system call. It uses the
GlobalAbbInfo interface to allow a later plug'n replace of the
RunningTask analysis, which is a data flow analysis by the symbolic
execution of the system state.

Until now only ActivateTask, TerminateTask and ChainTask are specialized
and scheduling is only then specialized when only one other task is a
possible next ABB.

Change-Id: Ie46b9a03fb8cdddb00ec6940bc8d5a491bf58455