generator: new pass SymbolicSystemExecution
The idea with symbolic system execution is to go all possible paths
through the system without merging the states. The result of the SSE is
a graph where system states are connected. Each system state exists only
once in that graph, so there will be cycles.

The goal of the SSE is to get a more precise picture of system
transitions than with the dataflow analysis of the system state, since
states are not merged, when more than one state flows into an
ABB. Downside: It might and probably will (at least with ISRs) explode.

Change-Id: Ie75217df26b048f9cc206391929b5c31ae58cd71