generator: new ABB kickoff
Since we now return in all cases to the computation block we were
interrupted by a sporadic event, we could no longer distinguish between
the start and the preemption of a subtask. A new block kickoff in front
of the first computatation block makes this possible. This will unbreak alarm1f

Change-Id: I234bcd9fc1404b5b513ab2b929dc33edfd0b84c8