generator: GenerateAssertionsPass
Assertions are generated from the global_abb_information. For the part
of the systemstate that is constant before and after a systemcall, we
generate a set of asserts. At the moment there are two types of asserts:

   TaskIsReady and TaskIsSuspended

which check whether a task is properly in suspended mode. There is also
a first optimization for avoid double checking of asserts during a
kernelphase.
                            +----------------+
                            |                |
                            |                |
  +------------------+      |      +---------V--------+
  |                  |      |      |                  |
  |  Before1         |      |      |   Before2        |
  |                  |      |      |                  |
  +-------+----------+      |      +---------+--------+
          |                 |                |
          |                 |                |
          |                 |                |
  +-------V----------+      |       +--------V--------+
  |                  |      |       |                 |
  | SysCall1         +------+       |  SysCall2       |
  |                  |              |                 |
  +------------------+              +--------+--------+
                                             |
                                             |
               +-----------------------------+
               |
               |
  +------------V-----+              +------------------+
  |                  |              |                  |
  | After 1          |              | After 2          |
  |                  |              |                  |
  +------------------+              +------------------+

In this flow graph (that shows global system flow edges) we see 2
system calling sites (Before1/Syscall1/After1 and
Before2/Syscall2/After2). Syscall1 dispatches not directly back to
After1. The situation is that for the backdispatch (from 2 to 1) the
kernespace consists of three blocks:

- Before 2: the before assertions of syscall 2
- Syscall2: the actual syscall
- After  1: the continuation point of Task1 with assertions

The optimization is, that we wipe out all assertions that are done
before all systemcalls that continue in After 1, since they would double
check something that is not nessecary.

Change-Id: I623198e0fc2a68a4188f594b8436431d6374a62a