generator/bugfix: missing activation edges to kickoffs
Since in the SSE irq edges go to the interrupt handler and not the
activated (or continued or whatever edge) we would miss edges that model
the effect of the interrupt. Therefore the system-level construction
pass no searches for all states that return to userland, when started
from userland. This models the same behaviour as the interrupt
transactions in the ISR2 handlers of the state flow.

Change-Id: Idfc83b0e4adbad8a3ab77144ff6040439878a55b