generator: DominanceAnalysis: find syscall immdom tree and regions
The dominator tree can be filtered to contain only system calls, the
dominance relation is not touched by this filtering. From the syscall
dominator tree we can identify all dominance regions of the system. A
dominance region is a region of blocks, where all children are dominated
by the root of the tree.

Change-Id: Ia385bfdaeef35ad8c4f8f7cd995d318da2c85019