arch/i386: save_sp as offset in stackptr* array
save_sp is now an index to a const stackptr* array stored in ROM.

The actual 16-bit index value is stored twice in save_sp to detect errors by
redundancy.

To detect if the IRQ interrupted an userspace task %cr3 was compared
to pagedir_os. Now, save_sp is checked to be >0.

Change-Id: Ia0eb41cef20a8d80ebded9d13fbda630d26940ac