x86: Booting other cpus
Booting a second cpu on an intel machine requires inter-processor
interrupts. For compatibility with older intel platforms, it is
necessary to find out if the APIC is an external chip or not. These
two requirements are met by adding the following to functions to the
LAPIC class:
    - trigger_IPI:      Triggers an ipi
    - is_IPI_pending:   Returns if the previous ipi was already
                        delivered or not
    - is_external_APIC: Returns if the APIC is an external chip

Additionally, read access to specific parts of the memory is required.
The MMU is therefore expanded with enable and dissable functions.

The architecture specific startup code calls boot_cpu with a stack
pointer which starts the second cpu by sending an asserting INIT IPI,
a deasserting INIT IPI, and if necessary up to three startup IPIs.
The entry point of the application cpu is at address 0x40000, where
the setup code is copied before triggering any ipi.

The setup code sets up a basic gdt and idt and switches to protected
mode. _mp_start prepares the environment for C++ code and then jumps
to it.

To be compatible with other platforms, the generic
start_hardware_thread function is introduced which can be implemented by
each architecture. It allows to start hardware threads with arbitary
functions.

Change-Id: I277d2a47e14570847e856038ef155cbf445fa562