arch/i386: place .rodata correctly
The "remaining" read-only data is placed in the remaining text region,
which is mapped readable by the linker script generator.
Until now, the rodata was placed after the remaining .text region and not
mapped by the generator, resulting in a working system only as long as it
still fitted into the _stext/_etext page.

Change-Id: I3496baa2df28d9cfd206112f6208cfbf7ca0cf84