CMake: Use gnu assembler for *.s
Introduces the ASM-ATT language in CMake. This affects all files
ending with ".s". Now, the gnu assembler is used to compile them.
This is required due to clang 3.4's lack of support for the
.code16 statement.

To preserve compatibility with the macros used in arch/i386/idt.S,
files ending with ".S" are still compiled with the clang assembler
(CMake language ASM) that was the default before this change.

Change-Id: Ia17dcd0fb7acc0f001eab38d61217e9a33ef7db0