Example app for CheckedObjects without the generator
Implements a new app that uses the previously implemented facilities
for CheckedObjects. Later, the implementation details should be
removed and the generator should handle inserting the CheckedObjects
into the dependability queue and the actual calling of the acquire
and release functions.

Currently, libdepsvc is linked into generic. In the future, this should
happen within the applications.

Change-Id: I048aa494e71fe849cea2cf71356114c071a797f1