Replace Dependability_Queue with Dependability_Scheduler
Simplifies the interface used to obtain the next CheckedObject that
shall be checked. This reduces pointer usage once more.

The usage of the queue is completely removed and instead a (currently
very simple) scheduler is added that just keeps track which is the next
index to be used in the array containing the CheckedObjects.

Change-Id: I53e92f5aa96f874f1222da26d597925a03d5c625