Make checksum calculation function changeable
Extends the example app by another CheckedObject that has its own
checksum calculation function.

The generator generates a multiplexer function that decides which
function to call for the checksum calculation. It is called by the
dependability service instead of a direct call of crc32.

crc32 is now made public to allow the application to use it in its own
checksum calculation function.

Change-Id: I24b84348a851a8dad71057d0b7611312855bb761