Introduce a new policy for the dependability scheduler
The new policy requires an timestamp for each CheckedObject and an
age counter that is incremented each time, get_next is called. The
method then finds the CheckedObject with the timestamp that is the
farest in the past and returns the reference to it. A premature check,
whether the CheckedObject is currently acquired, is also made.

A successful update issues the age counter to be written to the
CheckedObject's timestamp (weight) field.

Change-Id: I31f8a5f6f808ac7d08d31ab88b87043b2e518bd4