scheduler: only use references to Task objects
If Task objects, which are constexpr, are passed as `const Task' and not
as `const Task &' the propagation of the constexpr information is not
ensured. In clang-3.4, this information is propagated only sometimes,
and I could not determine the rules of propagation.

This change *WILL* change as well the encoded and the unencoded
binary. Only to give an impression of the changes in terms of code-size
for bench-coptermock:

unencoded(base): 301400 -> 296468 (bytes)
encoded(base):   360931 -> 344547 (bytes)

Change-Id: I81cc14b363921c602446243d10e80d704e052787