SSE: Add save state graph before cutting out the ISRs
We save the state graph, before we cut out the ISRs. This allows us to
access the full state graph with the SavedStateTransition edges
afterwards. We need this access to construct a finite state machine
afterwards.

Also a bug was fixed. In some rare circumstances a node in the state
graph was not part of self.states. So the references state node was
floating around.

Change-Id: I3a0de4b22ac21f8ae73ae6f67f91daa00ac395e8