Scheduler: Simplify the scheduler
Both of our scheduler implementations took a very long time to
compile. Therefore we compile the Reschedule function now to a single
function, which is referenced from many locations. This will result in
smaller system calls, but less redundancy, and scheduler refinement is
not possible anymore.

In the encoded case, we had to simplify the updateMax cascade, and
insert an recode of the sigId variable, because with a large number of
tasks, these signatures did overflow.

Change-Id: I28daa9a0a8e9ee0574f2c3ac6968c6b904a69609