state-flow: make analysis more efficient
We do not spawn a sub-flow analysis for every computation block, for
every sporadic event, but we calculate the influence of the sporadic
event once and "mask" it onto the incomming system state when triggered
the sporadic event.

This is possible, because not all system calls are allowed in ISR2s.

Change-Id: I2d7e04d37039e663e7838207ab6e9d12356ad4d7