state-asserts: more dense assertions
The state asserts now use a single conditional. For all values that
should be equal we OR the check results and check whether the ORed
combination is zero. For check that should be unequal zero, we calculate
the product of all results and check for unequal zero.

Additionally, we limit the number of assertions per system call to 50.

Change-Id: I066e6be7ef8bd40636bf943f5524e4ea7d01deaf