Event Support: Readout from application
We support now events in the application! The user can define events in
the OIL description. We generate system objects for the event, while we
already decoupling event names. We generate a system object for
each (Subtask, Event) tuple.

The arguments of SetEvent, GetEvent, ClearEvent, WaitEvent are reduced
to lists of system-object events.

The middle-end, and back-end do not support events at all. Therefore,
the test cases will fail.

Change-Id: I051a81d76ecebd379a2f38f03522797494513729