Event Support: Bugfix for non-preemptable tasks
This was a hard one! When non preemptable tasks do an WaitEvent and
_block_ they fall back to their original priority. In the unencoded
dOSEK this was already implemented. But for the encoded dOSEK we had to
implement a encoded path. It is now done with a little bit of ANBD
magic. See os/scheduler/tasklist.h.in for the detailed documentation.

Change-Id: Ie5e894030294503b4afed0364c4340e1dda6e34c