ABB merge: use dominance/postdominance information
In order to cope better with large complex applications, like KESO
applications, we implement an additional merging strategy: use dominance
and post-dominance relationship.

The immediate dominator and an post-dominator form an single-entry
single-exit region for an ABB. (I tried to implement the linear time
johnson PLDI'94 algorithm, but failed..) So if this SESE region is
mergable, we can simply merge it!

Result on keso_gc_coffebreak:
  SSE states before: 47802
  SSE states after: 30064

Change-Id: I8c58fcb5d0df4053d9dbf1cc2b055d32663e4207