toolchain: Invoke toolchain with cross compiler
We actually need a cross compiler setup, even for the i386 variant.
This change invokes the i386-elf- toolchain.
(It only worked in Linux, as it uses the same ELF format for executables)

The corresponding tools (nm, objdump) are propagated to the various
scripts and builds.

Fixes for Linux variant.
Fixes for posix variant.
Use /proj/i4danceos/tools/llvm-3.4 as Hint

Change-Id: I46b9dcdb5e0292b4083a0a6456384ef71f8ebe31