config: distributed constraints
constraints on the configuration can now be cluttered over the whole
(python) source code. The constraints are at the moment checked at
configuration time of the system, and not in every generator call.

As an example, you can now use

 config-constraint-: dependability.state-asserts -> (os.systemcalls == normal)

All constraints are started with `config-constraint-:'. The constraint
can be continued to the following line with ':-:'. For example,

  config-constraint-: os.specialize ->
               :-:      (arch.self == ARM)

Change-Id: I84dda39ccaa862b9dc7d4b2b53fe205ee2b8f93e