basic tasks: Implementation for posix
Basic tasks are a little bit complicated to implement. If we allow basic
tasks to run on a shared stack, we have to ensure that the stack is
switched correctly for all combinations of BT->ET, ET->BT, BT->BT task
switches. The testcases in app/ecc1/bt1 hopefully ensure the correct
implementation.

Change-Id: I35ad89e0d02b83383cc6637b65be283a7436a2bc