encoded/events: check calculated task priority before usage
The priority of a task is caluclated from its dynamic-priority state
word and the event state words. Until now, we did not check this value
for being a correctly encoded value, which causes quite a number of SDCs
in the bench-coptermock-isorc numbers. Therefore, we check this value
for validity and hope for the best.

Change-Id: Ie95c1b64b56d79a6110455270fb099fee919269f